Exract and fix private metadata tests
